move taxonomy to its own script file
