+++ tools/tagger/tagger.js
@@ -1,83 +1,5 @@
-import ffetch from '../../scripts/ffetch.js';
 import { createTag, getPalette } from '../../libs/utils/utils.js';
-
-// export async function getIndexDataByPath(name) {
-function titleToName(name) {
-  return name.toLowerCase().replace(' ', '-');
-}
-
-const taxonomyEndpoint = '/tools/taxonomy.json';
-let taxonomyPromise;
-function fetchTaxonomy() {
-  if (!taxonomyPromise) {
-    taxonomyPromise = new Promise((resolve, reject) => {
-      (async () => {
-        try {
-          const taxonomyJson = await ffetch(taxonomyEndpoint).sheet('tags').all();
-          const taxonomy = {};
-          let curType;
-          let l1;
-          taxonomyJson.forEach((row) => {
-            if (row.Type) {
-              curType = row.Type;
-              taxonomy[curType] = {
-                title: curType,
-                name: titleToName(curType),
-                path: titleToName(curType),
-                hide: row.hide,
-              };
-            }
-
-            if (row['Level 1']) {
-              l1 = row['Level 1'];
-              taxonomy[curType][l1] = {
-                title: l1,
-                name: titleToName(l1),
-                path: `${titleToName(curType)}/${titleToName(l1)}`,
-                hide: row.hide,
-              };
-            }
-          });
-          resolve(taxonomy);
-        } catch (e) {
-          reject(e);
-        }
-      })();
-    });
-  }
-
-  return taxonomyPromise;
-}
-
-const getDeepNestedObject = (obj, filter) => Object.entries(obj)
-  .reduce((acc, [key, value]) => {
-    let result = [];
-    if (key === filter) {
-      result = acc.concat(value);
-    } else if (typeof value === 'object') {
-      result = acc.concat(getDeepNestedObject(value, filter));
-    } else {
-      result = acc;
-    }
-    return result;
-  }, []);
-
-/**
- * Get the taxonomy of a hierarchical json object
- * @returns {Promise} the taxonomy
- */
-export function getTaxonomy() {
-  return fetchTaxonomy();
-}
-
-/**
- * Returns a taxonomy category as an array of objects
- * @param {*} category
- */
-export const getTaxonomyCategory = async (category) => {
-  const taxonomy = await getTaxonomy();
-  return getDeepNestedObject(taxonomy, category)[0];
-};
+import { getTaxonomy } from '../../scripts/taxonomy.js';
 
 let selectedOrder = [];
 
@@ -114,6 +36,7 @@ function initTaxonomy(taxonomy) {
     html += '</div>';
   });
   const results = document.getElementById('results');
+  if (!results) return;
   results.innerHTML = html;
 }
